+++ src/bytecodes/Sub.cpp
@@ -22,7 +22,7 @@ void Sub::execute(){
         float float2 = Program::runtime_stack[Program::stack_pointer]->getFloat();
         Program::runtime_stack.pop_back();
         Program::runtime_stack.pop_back();
-        Program::runtime_stack.push_back((new Value(float1 - float2)));
+        Program::runtime_stack.push_back((new Value((float) (float1 - float2))));
     }
     else if (type == "short"){
         short short1 = Program::runtime_stack[Program::stack_pointer - 1]->getShort();
